\docType{package}
\name{ PGAS-package}
\alias{PGAS-package}
\title{Implementation of  Particle Gibbs Particle Gibbs Method and its Extensions and Variants}

\description{
Implementation of  Particle Gibbs Particle Gibbs Method and its Extensions and Variants
}

\details{
Implementation of  Particle Gibbs Particle Gibbs Method and its Extensions and Variants

}

\references{
[1] Lindsten, Fredrik, Michael I. Jordan, and Thomas B. Schön. "Particle Gibbs with ancestor sampling." The Journal of Machine Learning Research 15.1 (2014): 2145-2184.

[2] C. Andrieu, A. Doucet and R. Holenstein, "Particle Markov chain Monte Carlo methods" Journal of the Royal Statistical Society: Series B, 2010, 72, 269-342.

[3] Rainforth, Tom, et al. "Interacting particle Markov chain Monte Carlo."
International Conference on Machine Learning. 2016.

[4] Svensson, Andreas, et al. "Computationally efficient Bayesian learning of
Gaussian process state space models." Artificial Intelligence and Statistics. 2016.
}
