\docType{package}
\name{An Introduction of the package PGAS}
\alias{package-PGAS}
\title{Implementation of  Particle Gibbs with Ancestor Sampling and  and the Particle Marginal Metropolis-Hastings algorithms}

\description{
Implementation of  Particle Gibbs with Ancestor Sampling ([1]) and the Particle Marginal Metropolis-Hastings ([2]) algorithms. See  (<http://www.it.uu.se/katalog/freli660/software>) for implementation in 'MATLAB'.
}

\details{

This package offeres a inverse gamma prior based solution for the PGAS and PMMH algorithms. In future, more generalized priors are planned to be supported. We refer to [1] for theoretical details of PGAS, [2] for details on PMMH algorithm and [3] for implementation details of PGAS and PMMH in MATLAB.

}

\references{
[1] Lindsten, Fredrik, Michael I. Jordan, and Thomas B. Schön. "Particle Gibbs with ancestor sampling." The Journal of Machine Learning Research 15.1 (2014): 2145-2184.

[2] C. Andrieu, A. Doucet and R. Holenstein, "Particle Markov chain Monte Carlo methods" Journal of the Royal Statistical Society: Series B, 2010, 72, 269-342.

[3] Lindsten, Fredrik, http://www.it.uu.se/katalog/freli660/software
}
